feat: 네트워크 로직 정리
