--- EzyBook/Data/Remote/Router/Interfaces/GetRouter.swift
@@ -1,11 +1,0 @@
-//
-//  GetRouter.swift
-//  EzyBook
-//
-//  Created by youngkyun park on 5/12/25.
-//
-
-import Foundation
-
-/// Get 전용 라우터
-protocol GetRouter: NetworkRouter { }
