+++ EzyBook/Infrastructure/Networking/DefaultNetworkService.swift
@@ -8,7 +8,7 @@
 import Foundation
 import Alamofire
 
-final class DefaultNetworkService: NetworkService, MultiPartNetworkSevice {
+final class DefaultNetworkService: NetworkService {
 
     private let decodingService: ResponseDecoder
     private let interceptor: TokenInterceptor? // 이것도 추상화를 해줘야하나
@@ -28,15 +28,34 @@ final class DefaultNetworkService: NetworkService, MultiPartNetworkSevice {
         } catch {
             throw APIError(localErrorType: .missingEndpoint)
         }
-    
-        let response = await session.request(urlRequest)
-            .validate(statusCode: 200...299)
-            .serializingData()
-            .response
+        //모든 분기 경로에서 초기화가 보장될 경우 초기화와 선언을 동시에 하지 않아도 괜찮음
+        let response: AFDataResponse<Data>
         
+        if let multipartForm = (router as? MultipartRouter)?.multipartFormData {
+            guard let endpoint = router.endpoint else {
+                throw APIError(localErrorType: .missingEndpoint)
+            }
+            
+            response = await session
+                .upload(multipartFormData: multipartForm,
+                        to: endpoint,
+                        method: router.method,
+                        headers: router.headers)
+                .validate(statusCode: 200...299)
+                .serializingData()
+                .response
+            
+        } else {
+            response = await session
+                      .request(urlRequest)
+                      .validate(statusCode: 200...299)
+                      .serializingData()
+                      .response
+        }
+        
+ 
         if let request = response.request {
             print("Full URL:", request.url?.absoluteString ?? "nil")
-            print("Full URL:", urlRequest)
             print("HTTP Method:", request.httpMethod ?? "nil")
             print("Headers:", request.headers)
         }
@@ -77,68 +96,6 @@ final class DefaultNetworkService: NetworkService, MultiPartNetworkSevice {
         }
     }
     
-    func fetchMultiPartData<T: Decodable & EntityConvertible, R: NetworkRouter>(dto: T.Type ,_ router: R) async throws -> T {
-        
-        let urlRequest: URLRequest
-        do {
-            urlRequest = try router.asURLRequest()
-        } catch {
-            throw APIError(localErrorType: .missingEndpoint)
-        }
-    
-        guard let multipartBuilder = router.multipartFormData else {
-               throw APIError(localErrorType: .missingEndpoint)//APIError(localErrorType: .invalidMultipart)
-           }
-        
-        let response = await session.upload(multipartFormData: multipartBuilder, to: router.endpoint!, method: router.method, headers: router.headers)
-            .validate(statusCode: 200...299)
-            .serializingData()
-            .response
-        
-        if let request = response.request {
-            print("Full URL:", request.url?.absoluteString ?? "nil")
-            print("Full URL:", urlRequest)
-            print("HTTP Method:", request.httpMethod ?? "nil")
-            print("Headers:", request.headers)
-            
-            
-        }
-        
-        #if DEBUG
-        let urlString = urlRequest.url?.absoluteString ?? "Invalid URL"
-        #endif
-        
-        let statusCode = response.response?.statusCode
-        
-        switch response.result {
-        case .success(let data):
-            
-            let decodedResult = decodingService.decode(data: data, type: dto)
-            
-            #if DEBUG
-            NetworkLog.success(url: urlString, statusCode: statusCode ?? 0, data: data)
-            #endif
-           
-            switch decodedResult {
-            case .success(let decodedDTO):
-                return decodedDTO
-            case .failure(let decodeError):
-                throw decodeError
-            }
-        case .failure(let afError):
-            #if DEBUG
-            NetworkLog.failure(url: urlString, statusCode: statusCode ?? 0, data: response.data)
-            #endif
-            
-            let responseData = response.data
-            if let code = statusCode {
-                throw APIError(statusCode: code, data: responseData)
-            } else {
-                let errorCode = (afError as NSError).code
-                throw APIError(statusCode: errorCode, data: responseData)
-            }
-        }
-    }
     
 }
